@@ -3,17 +3,17 @@ description: A new Flutter package project.
 version: 0.0.1
 
 environment:
-  sdk: ">=2.17.1 <3.0.0"
+  sdk: ">=2.19.2 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  bloc: ^8.1.1
+  bloc: ^7.1.0
   equatable: ^2.0.5
   intl: ^0.17.0
-  flutter_bloc: ^8.1.1
+  flutter_bloc: ^7.2.0
 
 dev_dependencies:
-  bloc_test: ^9.1.1
+  bloc_test: ^8.1.0
   flutter_test:
     sdk: flutter
