@@ -1,5 +1,3 @@
-// @dart=2.9
-
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:invalid/invalid.dart';
@@ -12,17 +10,17 @@ class ValidationForm<KeyType> extends StatelessWidget {
   /// Will be called once the form turns from invalid to valid.
   /// But be aware, fields (therefor also validators) can be removed asynchronously.
   /// When you remove all fields (therefor also validators) the form turns valid and this callback is called.
-  final Function(FormValidationState<KeyType>) onFormTurnedValid;
+  final Function(FormValidationState<KeyType>)? onFormTurnedValid;
 
   /// Will be called once the form turns from valid to invalid.
   /// But be aware, fields (therefor also validators) are added asynchronously (for example in initState of TextField),
   /// which means, that the form is valid at the beginning and once you add fields it can turn invalid.
   /// That's why it can happen that this callback is called immediately, after adding fields.
-  final Function(FormValidationState<KeyType>) onFormTurnedInValid;
+  final Function(FormValidationState<KeyType>)? onFormTurnedInValid;
 
   /// [onUpdate] will be called when the state of [FormValidationState] changes.
-  final Function(FormValidationState<KeyType>) onUpdate;
-  final Function(FormValidationCubit<KeyType>) onFormValidationCubitCreated;
+  final Function(FormValidationState<KeyType>)? onUpdate;
+  final Function(FormValidationCubit<KeyType>)? onFormValidationCubitCreated;
 
   /// If true, the validation messages are shown immediately, if false it must be enabled before.
   /// Use case for this is when you have a form, and you want validation messages to appear after the user
@@ -31,10 +29,10 @@ class ValidationForm<KeyType> extends StatelessWidget {
   final Widget child;
 
   const ValidationForm(
-      {Key key,
-      @required this.child,
+      {Key? key,
+      required this.child,
       this.enabled = false,
-      this.formValidators,
+      this.formValidators = const [],
       this.onUpdate,
       this.onFormTurnedValid,
       this.onFormTurnedInValid,
@@ -79,7 +77,7 @@ extension BuildContextExtensions on BuildContext {
 }
 
 class ValidationCapability<KeyType> {
-  FormValidationCubit<KeyType> _formValidationBloc;
+  late FormValidationCubit<KeyType> _formValidationBloc;
   KeyType validationKey;
   List<FieldValidator<dynamic, KeyType, dynamic>> validators;
   String fieldName;
@@ -89,8 +87,8 @@ class ValidationCapability<KeyType> {
   }
 
   ValidationCapability(
-      {@required this.validationKey,
-      this.fieldName,
+      {required this.validationKey,
+      this.fieldName = '',
       Iterable<FieldValidator<dynamic, KeyType, dynamic>> validators = const []})
       : validators = validators.toList();
 
@@ -100,24 +98,17 @@ class ValidationCapability<KeyType> {
 
   FormValidationCubit<KeyType> _addFieldAndGetForm(
       Iterable<FieldValidator<dynamic, KeyType, dynamic>> validators, KeyType key, BuildContext context,
-      {String fieldName}) {
-    if (key == null) return null;
+      {String? fieldName}) {
     var formValidationBloc = BlocProvider.of<FormValidationCubit<KeyType>>(context);
     formValidationBloc.addOrReplaceField(field);
     return formValidationBloc;
   }
 
   void updateFieldValue(dynamic newVal) {
-    if (_formValidationBloc == null) {
-      throw UninitializedException<KeyType>();
-    }
     _formValidationBloc.updateFieldValue(validationKey, newVal);
   }
 
   void updateField() {
-    if (_formValidationBloc == null) {
-      throw UninitializedException<KeyType>();
-    }
     _formValidationBloc.updateFieldValue(validationKey, field);
   }
 
@@ -135,22 +126,21 @@ class ValidationCapability<KeyType> {
 }
 
 class TextValidationCapability<KeyType> extends ValidationCapability<KeyType> {
-  TextEditingController _controller;
-  bool autoValidate;
+  late TextEditingController _controller;
+  final bool autoValidate;
 
-  TextValidationCapability(
-      {@required KeyType validationKey,
-      String fieldName,
-      Iterable<FieldValidator<dynamic, KeyType, dynamic>> validators = const [],
-      this.autoValidate = true})
-      : super(validationKey: validationKey, validators: validators, fieldName: fieldName);
+  TextValidationCapability({
+    required KeyType validationKey,
+    String fieldName = '',
+    Iterable<FieldValidator<dynamic, KeyType, dynamic>> validators = const [],
+    this.autoValidate = true,
+  }) : super(validationKey: validationKey, validators: validators, fieldName: fieldName);
 
-  @override
-  void init(BuildContext context, {@required TextEditingController controller}) {
+  void initWithTextEditingController(BuildContext context, {required TextEditingController controller}) {
     super.init(context);
     _controller = controller;
     updateFieldValue(_controller.text);
-    if (autoValidate) _controller?.addListener(() => updateFieldValue(_controller.text));
+    if (autoValidate) _controller.addListener(() => updateFieldValue(_controller.text));
   }
 
   @override
@@ -163,12 +153,4 @@ class TextValidationCapability<KeyType> extends ValidationCapability<KeyType> {
 
   @override
   int get hashCode => super.hashCode ^ _controller.hashCode ^ autoValidate.hashCode;
-}
-
-class UninitializedException<KeyType> implements Exception {
-  @override
-  String toString() {
-    return "UninitializedException: ValidatorCapability was used before calling init(BuildContext context) "
-        "or was not initialized inside a context which contains a FormValidationBloc of type ${KeyType.runtimeType}";
-  }
 }
